[BEAM-4294] [BEAM-4360] Join translation and ReduceByKey test suite where moved to org.apache.beam.* package. Small imports fixes after rebase to seznam/beam eff3ffd5d224c6628606b355e3a5b6d8bc40150b.
