+++ sdks/java/extensions/euphoria/euphoria-beam/src/main/java/org/apache/beam/sdk/extensions/euphoria/beam/FlowTranslator.java
@@ -17,7 +17,6 @@
  */
 package org.apache.beam.sdk.extensions.euphoria.beam;
 
-import cz.seznam.euphoria.beam.JoinTranslator;
 import java.util.IdentityHashMap;
 import java.util.Map;
 import org.apache.beam.sdk.Pipeline;
