+++ sdks/java/extensions/euphoria/euphoria-beam/src/main/java/org/apache/beam/sdk/extensions/euphoria/beam/JoinTranslator.java
@@ -1,21 +1,21 @@
-package cz.seznam.euphoria.beam;
+package org.apache.beam.sdk.extensions.euphoria.beam;
 
-import cz.seznam.euphoria.beam.common.InputToKvDoFn;
-import cz.seznam.euphoria.beam.io.KryoCoder;
-import cz.seznam.euphoria.beam.join.FullJoinFn;
-import cz.seznam.euphoria.beam.join.InnerJoinFn;
-import cz.seznam.euphoria.beam.join.JoinFn;
-import cz.seznam.euphoria.beam.join.LeftOuterJoinFn;
-import cz.seznam.euphoria.beam.join.RightOuterJoinFn;
-import cz.seznam.euphoria.beam.window.WindowingUtils;
-import cz.seznam.euphoria.core.client.dataset.windowing.Window;
-import cz.seznam.euphoria.core.client.functional.BinaryFunctor;
-import cz.seznam.euphoria.core.client.functional.UnaryFunction;
-import cz.seznam.euphoria.core.client.operator.Join;
-import cz.seznam.euphoria.core.client.util.Pair;
+import org.apache.beam.sdk.extensions.euphoria.beam.common.InputToKvDoFn;
+import org.apache.beam.sdk.extensions.euphoria.beam.join.FullJoinFn;
+import org.apache.beam.sdk.extensions.euphoria.beam.join.InnerJoinFn;
+import org.apache.beam.sdk.extensions.euphoria.beam.join.JoinFn;
+import org.apache.beam.sdk.extensions.euphoria.beam.join.LeftOuterJoinFn;
+import org.apache.beam.sdk.extensions.euphoria.beam.join.RightOuterJoinFn;
+import org.apache.beam.sdk.extensions.euphoria.beam.window.WindowingUtils;
 import java.util.List;
 import org.apache.beam.sdk.coders.Coder;
 import org.apache.beam.sdk.coders.KvCoder;
+import org.apache.beam.sdk.extensions.euphoria.beam.io.KryoCoder;
+import org.apache.beam.sdk.extensions.euphoria.core.client.dataset.windowing.Window;
+import org.apache.beam.sdk.extensions.euphoria.core.client.functional.BinaryFunctor;
+import org.apache.beam.sdk.extensions.euphoria.core.client.functional.UnaryFunction;
+import org.apache.beam.sdk.extensions.euphoria.core.client.operator.Join;
+import org.apache.beam.sdk.extensions.euphoria.core.client.util.Pair;
 import org.apache.beam.sdk.transforms.ParDo;
 import org.apache.beam.sdk.transforms.join.CoGbkResult;
 import org.apache.beam.sdk.transforms.join.CoGroupByKey;
