+++ sdks/java/extensions/euphoria/euphoria-beam/src/main/java/org/apache/beam/sdk/extensions/euphoria/beam/common/package-info.java
@@ -15,40 +15,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package cz.seznam.euphoria.beam.window;
-
-import java.util.Objects;
-import org.apache.beam.sdk.extensions.euphoria.core.client.dataset.windowing.Window;
-
 
 /**
- * Window used as type parameter of {@link BeamWindowing}.
+ * A set of commonly used classes enabling some code reuse.
  */
-final class UnsupportedWindow extends Window<UnsupportedWindow> {
-
-  private UnsupportedWindow(){
-    //Do not instantiate
-  }
-
-  @Override
-  public int compareTo(UnsupportedWindow o) {
-    throw new UnsupportedOperationException();
-  }
-
-  @Override
-  public long maxTimestamp() {
-    throw new UnsupportedOperationException();
-  }
-
-  @Override
-  public int hashCode() {
-    throw new UnsupportedOperationException();
-  }
-
-  @Override
-  public boolean equals(Object obj) {
-    return Objects.equals(this, obj);
-  }
-
-
-}
+package org.apache.beam.sdk.extensions.euphoria.beam.common;
