+++ sdks/java/extensions/euphoria/euphoria-beam/src/main/java/org/apache/beam/sdk/extensions/euphoria/beam/join/LeftOuterJoinFn.java
@@ -1,18 +1,19 @@
-package cz.seznam.euphoria.beam.join;
+package org.apache.beam.sdk.extensions.euphoria.beam.join;
 
-import cz.seznam.euphoria.beam.SingleValueCollector;
-import cz.seznam.euphoria.core.client.functional.BinaryFunctor;
-import cz.seznam.euphoria.core.client.util.Pair;
+import org.apache.beam.sdk.extensions.euphoria.beam.SingleValueCollector;
+import org.apache.beam.sdk.extensions.euphoria.core.client.functional.BinaryFunctor;
+import org.apache.beam.sdk.extensions.euphoria.core.client.util.Pair;
 import org.apache.beam.sdk.transforms.join.CoGbkResult;
 import org.apache.beam.sdk.values.KV;
 import org.apache.beam.sdk.values.TupleTag;
 
-/**
- * Right outer join implementation of {@link JoinFn}.
- */
-public class RightOuterJoinFn<LeftT, RightT, K, OutputT> extends JoinFn<LeftT, RightT, K, OutputT> {
 
-  public RightOuterJoinFn(
+/**
+ * Left outer join implementation of {@link JoinFn}.
+ */
+public class LeftOuterJoinFn<LeftT, RightT, K, OutputT> extends JoinFn<LeftT, RightT, K, OutputT> {
+
+  public LeftOuterJoinFn(
       BinaryFunctor<LeftT, RightT, OutputT> joiner,
       TupleTag<LeftT> leftTag,
       TupleTag<RightT> rightTag) {
@@ -31,22 +32,24 @@ public class RightOuterJoinFn<LeftT, RightT, K, OutputT> extends JoinFn<LeftT, R
 
     SingleValueCollector<OutputT> outCollector = new SingleValueCollector<>();
 
-    for (RightT rightValue : rightSIdeIter) {
-      if (leftSideIter.iterator().hasNext()) {
-        for (LeftT leftValue : leftSideIter) {
+    for (LeftT leftValue : leftSideIter) {
+      if (rightSIdeIter.iterator().hasNext()) {
+        for (RightT rightValue : rightSIdeIter) {
           joiner.apply(leftValue, rightValue, outCollector);
           c.output(Pair.of(key, outCollector.get()));
         }
       } else {
-        joiner.apply(null, rightValue, outCollector);
-        c.output(Pair.of(key, outCollector.get()));
+          joiner.apply(leftValue, null, outCollector);
+          c.output(Pair.of(key, outCollector.get()));
       }
     }
+
 
   }
 
   @Override
   public String getFnName() {
-    return "::right-outer-join";
+    return "::left-outer-join";
   }
+
 }
